+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/generators/AssistedFactoryImplFirGenerator.kt
@@ -8,6 +8,7 @@ import dev.zacsweers.metro.compiler.fir.Keys
 import dev.zacsweers.metro.compiler.fir.abstractFunctions
 import dev.zacsweers.metro.compiler.fir.classIds
 import dev.zacsweers.metro.compiler.fir.constructType
+import dev.zacsweers.metro.compiler.fir.copyTypeParametersFrom
 import dev.zacsweers.metro.compiler.fir.hasOrigin
 import dev.zacsweers.metro.compiler.fir.isAnnotatedWithAny
 import dev.zacsweers.metro.compiler.fir.predicates
@@ -135,14 +136,7 @@ internal class AssistedFactoryImplFirGenerator(session: FirSession) :
       Symbols.Names.MetroImpl -> {
         // TODO if there's no assisted params, we could optimize this to just be an object?
         createNestedClass(owner, name, Keys.AssistedFactoryImplClassDeclaration) {
-            for (typeParam in owner.typeParameterSymbols) {
-              typeParameter(typeParam.name, typeParam.variance, key = Keys.Default) {
-                if (typeParam.isBound) {
-                  typeParam.resolvedBounds.forEach { bound -> bound(bound.coneType) }
-                }
-              }
-            }
-
+            copyTypeParametersFrom(owner, session)
             superType(owner::constructType)
           }
           .symbol